Add production of xsdir/xsdata files for Serpent
Added a script allowing :
* the merging of the individual xsdir files,
* to give different Monte Carlo identifiers to isotopes with the same
  ZAID, such as metastable isotopes (resolves #1 issue) or TMC samplings
  of identical isotopes. These identifiers are used both in the merged
  xsdir file and in the header of the ACE files modified for this
  purpose.

Also, the Serpent script converting xsdir files to xsdata has been
improved, to allow it to convert more diverse identifiers (including
those described above).

As a result of these changes, all isotopes have the same name in the
Draglib and in the xsdata file. So they can be called the same way in
Dragon and Serpent.