Fix XSDATA metastable isotopes ZAID for Serpent2
This issue was caused by the script "xsdirconvert.pl" from VTT, that
incorrectly recognized metastable (isomeric) isotopes. Isotopes like
Pm148 and Pm148m were receiving the same ZAID identifier, leading
Serpent to incorrectly use the same cross sections for both.

This has now been corrected, by feeding this script with the
conventions it expects: the third digit of the ZAID is equal
to 3 for metastable isotopes. Therefore, it closes #10.

The induced error is non-negligible : see issue #10.